data-pipeline/requirements.txt
# Data Pipeline Dependencies
arxiv==1.4.8
scholarly==1.7.11
PyGithub==1.59.1
scrapy==2.11.0
apache-airflow==2.7.3

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.0
pymupdf==1.23.8

# NLP & Text Processing
spacy==3.7.2
transformers==4.35.2
sentence-transformers==2.2.2
nltk==3.8.1

# Data Processing
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2

# HTTP & Async
aiohttp==3.9.1
httpx==0.25.2
asyncio-throttle==1.0.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
structlog==23.2.0